Modify Medication Component Amount to allow decimal values

+++ trunk/CGTA_Input_Tester/ConverterLibrary/src/main/java/ca/cgta/input/model/inner/MedicationComponent.java
@@ -25,7 +25,7 @@ public class MedicationComponent {
     
     public String myComponentType;
     public Ce myComponentCode;
-    public int myComponentAmount;    
+    public double myComponentAmount;    
     public Ce myComponentUnits;
